Added paging.

+++ mkm/src/main/kotlin/mkm/repos/ExpansionRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface ExpansionRepository: CrudRepository<Expansion, Int>
+interface ExpansionRepository: CrudRepository<Expansion, Long>
